Refactor MainActivity setup
