+++ app/src/main/java/com/example/jellyfinandroid/utils/ImageLoaderInitializer.kt
@@ -1,0 +1,33 @@
+package com.example.jellyfinandroid.utils
+
+import android.content.Context
+import coil.Coil
+import coil.ImageLoader
+import coil.disk.DiskCache
+import coil.memory.MemoryCache
+import dagger.hilt.android.qualifiers.ApplicationContext
+import javax.inject.Inject
+import javax.inject.Singleton
+
+@Singleton
+class ImageLoaderInitializer @Inject constructor(
+    @ApplicationContext private val context: Context
+) {
+    fun initialize() {
+        val imageLoader = ImageLoader.Builder(context)
+            .memoryCache {
+                MemoryCache.Builder(context)
+                    .maxSizePercent(0.25)
+                    .build()
+            }
+            .diskCache {
+                DiskCache.Builder()
+                    .directory(context.cacheDir.resolve("image_cache"))
+                    .maxSizeBytes(50 * 1024 * 1024)
+                    .build()
+            }
+            .respectCacheHeaders(false)
+            .build()
+        Coil.setImageLoader(imageLoader)
+    }
+}
